+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - use [codecov](https://codecov.io) for code coverage
 - renamed command `odoo-helper print_config` to `odoo-helper print-config`
 - Added `odoo-helper test --coverage-skip-covered` option
+- Added `odoo-helper addons update-py-deps` command
 - Removed old unused options
   - `odoo-helper --addons-dir <addons_directory>`
   - `odoo-helper --downloads-dir <downloads_directory>`